move get_stacked_metadata into shared.utils
