+++ external/fv3fit/fv3fit/keras/_models/shared/utils.py
@@ -1,8 +1,28 @@
 from fv3fit._shared.packer import ArrayPacker
 import tensorflow as tf
-from typing import List, Optional, Sequence, Type
+from typing import List, Optional, Sequence, Type, Tuple, Dict, Any
+import xarray as xr
 from fv3fit.emulation.layers import StandardNormLayer, StandardDenormLayer, NormLayer
+from fv3fit._shared.stacking import stack
+
 import numpy as np
+
+
+def get_stacked_metadata(
+    names, ds: xr.Dataset, unstacked_dims: Sequence[str]
+) -> Tuple[Dict[str, Any], ...]:
+    """
+    Retrieve xarray metadata for dataset after stacking.
+
+    Returns a dict containing "dims" and "units" for each name.
+    """
+    ds = stack(ds, unstacked_dims=unstacked_dims)
+    metadata = []
+    for name in names:
+        metadata.append(
+            {"dims": ds[name].dims, "units": ds[name].attrs.get("units", "unknown")}
+        )
+    return tuple(metadata)
 
 
 def get_input_vector(
